CameraHMR/requirements_nover.txt
pytorch-lightning
smplx
pyrender
opencv-python
yacs
scikit-image
einops
timm
pandas
loguru
numpy
pyrootutils
rich
hydra-colorlog
albumentations
chumpy
flatten_dict
jpeg4py
joblib
